pyproject.toml
[tool.poetry]
name = "iris"
version = "0.22.1"
description = "Resilient Internet-scale measurement system."
license = "MIT"
authors = [
    "Matthieu Gouel <matthieu.gouel@lip6.fr>",
    "Maxime Mouchet <maxime.mouchet@lip6.fr>"
]
readme = "README.md"
homepage = "https://github.com/dioptra-io/iris"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Topic :: Internet",
]

[tool.poetry.dependencies]
python = "^3.8"
aioboto3 = "^9.2.0"
aiofiles = "^0.8.0"
aioredis = "^2.0.0"
aiosmtplib = "^1.1.6"
aiosqlite = "^0.17.0"
asgiref = {optional = true, version = "^3.2.7"}
async-timeout = "^4.0.1"
diamond-miner = "^0.5.0"
dramatiq = {extras = ["redis", "watch"], optional = true, version = "^1.12.1"}
email-validator = "^1.1.3"
fastapi = {optional = true, version = "^0.70.0"}
fastapi-users = {extras = ["sqlalchemy", "oauth"], version = "^8.1.2"}
# <0.21 is a condition of fastapi-users oauth plugin
httpx = "<0.21"
gunicorn = {optional = true, version = "^20.1.0"}
pycaracal = {optional = true, version = "^0.6.2"}
pydantic = "^1.8.2"
python-jose = {extras = ["cryptography"], optional = true, version = "^3.3.0"}
python-multipart = {optional = true, version = "^0.0.5"}
pytricia = {optional = true, version = "^1.0.2"}
SQLAlchemy-Utils = "^0.37.8"
sqlmodel = "^0.0.6"
starlette_exporter = {optional = true, version = "^0.11.0"}
tenacity = "^8.0.1"
uvicorn = {optional = true, version = "^0.16.0"}
zstandard = "^0.16.0"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
boto3-stubs = "^1.20.26"
bumpversion = "^0.6.0"
coverage = "^6.2"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.930"
pre-commit = "^2.16.0"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
pytest-cov = "^3.0.0"
pytest-clarity = "^1.0.1"
types-aiofiles = "^0.7.2"
alembic = "^1.7.5"

[tool.poetry.extras]
agent = [
    "pycaracal",
]
api = [
    "asgiref",
    "dramatiq",
    "fastapi",
    "gunicorn",
    "python-jose",
    "python-multipart",
    "starlette_exporter",
    "uvicorn"
]
worker = [
    "dramatiq",
    "pytricia"
]

[tool.pytest.ini_options]
# TODO: Make this compatible with PyCharm debugger?
addopts = "--doctest-modules --verbosity=2"
#addopts = "--ignore volumes --cov=iris --cov-report= --doctest-modules -vvs"
filterwarnings = [
    "ignore::DeprecationWarning:aioboto3.*:",
    "ignore::DeprecationWarning:aiofiles.*:",
    "ignore::DeprecationWarning:aiohttp.*:",
    "ignore::DeprecationWarning:urllib3.*:",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
